+++ apps/block_scout_web/assets/js/lib/market_history_chart.js
@@ -4,6 +4,7 @@ import humps from 'humps'
 import numeral from 'numeral'
 import { formatUsdValue } from '../lib/currency'
 import sassVariables from '../../css/app.scss'
+import { showLoader } from '../lib/utils'
 
 const config = {
   type: 'line',
@@ -129,6 +130,10 @@ class MarketHistoryChart {
 export function createMarketHistoryChart (el) {
   const dataPath = el.dataset.market_history_chart_path
   const $chartLoading = $('[data-chart-loading-message]')
+
+  const isTimeout = true
+  const timeoutID = showLoader(isTimeout, $chartLoading)
+
   const $chartError = $('[data-chart-error-message]')
   const chart = new MarketHistoryChart(el, 0, [])
   $.getJSON(dataPath, {type: 'JSON'})
@@ -143,6 +148,7 @@ export function createMarketHistoryChart (el) {
     })
     .always(() => {
       $chartLoading.hide()
+      clearTimeout(timeoutID)
     })
   return chart
 }